Refactoring of level loading
[level.cpp / level.h]
- Move level loading code to own file
- convert .info file format to json
- use nlohmann/json for parsing
- load json files
- use class-based approach in level

[assets.cpp /level.h]
- move asset loading to seperate file

[client.cpp / client.h]
- clean up asset/level loading a bit

[main.cpp]
- move some more global state

[net.cpp]
- use Uint16

[other.cpp]
- remove LoadLevelFile

[types.h]
- move some global state to level.h / assets.h / server.h

+++ source/net.cpp
@@ -101,7 +101,7 @@ void Buffer::Clear() {
 }
 
 // Seek the position pointer to an index.
-bool Buffer::Seek(uint index) {
+bool Buffer::Seek(Uint16 index) {
     if (index >= this->length) {
         return false;
     }
@@ -233,7 +233,7 @@ std::string& Buffer::ReadString() {
 }
 
 // Decrease the position by n. Will not increase beyond position 0
-void Buffer::DecreasePosition(uint n) {
+void Buffer::DecreasePosition(Uint16 n) {
     if( n >= this->length ) {
         this->Clear();
     } else {
